fix loader issue when swapping pages
