+++ src/index.js
@@ -7,7 +7,6 @@ import { BrowserRouter } from "react-router-dom";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
-
 	<React.StrictMode>
 		<BrowserRouter>
 			<App />
