+++ src/styles/App.css
@@ -632,6 +632,8 @@ footer a {
 .questions-page nav {
   text-align: left;
   padding-top: 20px;
+  position: relative;
+  left: 24px;
 }
 .questions-page nav a {
   font-size: 36px;
@@ -646,7 +648,7 @@ footer a {
 .questions-page ul {
   font-family: "Montserrat", sans-serif;
   max-width: 620px;
-  margin: 33px auto;
+  margin: 33px auto 43px auto;
 }
 .questions-page ul li {
   font-family: "Montserrat", sans-serif;
@@ -924,6 +926,7 @@ footer a {
   }
   .questions-page div nav {
     text-align: center;
+    position: static;
   }
   .outer-map-page div .home-icon-map-page {
     width: 100%;
