+++ src/styles/sass/partials/components/_QuestionsPage.scss
@@ -45,11 +45,12 @@
 	nav {
 		text-align: left;
 		padding-top: 20px;
+		position: relative;
+		left: 24px;
 
 		a {
 			font-size: $fs-600;
 			color: $clr-white;
-
 			&:hover,
 			&:focus {
 				color: $clr-teal;
@@ -64,7 +65,7 @@
 	ul {
 		font-family: $ff-secondary;
 		max-width: 620px;
-		margin: 33px auto;
+		margin: 33px auto 43px auto;
 
 		li {
 			font-family: $ff-secondary;
